Class/model fixes
